+++ tests/test_show.py
@@ -29,29 +29,22 @@ def test_show(sample_df, capsys):
         show_stats(sample_df, "NONSENSE")
 
 
-def test_namespace(sample_df, capsys):
-    sample_df.stats.show()
+def test_show_subsets(sample_df, capsys):
+    show_stats(sample_df.select("U", "int_col"))
     captured = capsys.readouterr()
     assert "Col (N=500)" in captured.out
-    assert "float_mean_2" in captured.out
-    assert "float_min_-7" in captured.out
-    sample_df.stats.show("cat")
+    assert "int_col" in captured.out
+
+    show_stats(sample_df.select("categorical_col"))
     captured = capsys.readouterr()
-    assert "Col (N=500)" in captured.out
-    assert "float_mean_2" not in captured.out
-    assert "float_min_-7" not in captured.out
-    assert "str_col" in captured.out
-    assert "enum_col" in captured.out
     assert "categorical_col" in captured.out
-    sample_df.select("U", "int_col").stats.show()
-    sample_df.select("categorical_col").stats.show()
 
 
 def test_show_empty(sample_df, capsys):
-    sample_df.select("U", "int_col").stats.show("cat")
+    show_stats(sample_df.select("U", "int_col"), "cat")
     captured = capsys.readouterr()
     assert captured.out == "No categorical columns found\n"
-    sample_df.select("str_col").stats.show("num")
+    show_stats(sample_df.select("str_col"), "num")
     captured = capsys.readouterr()
     assert captured.out == "No numerical columns found\n"